ds-nn-m5/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Document Processing
python-docx==1.1.0
PyMuPDF==1.23.26  # For PDF processing

# AWS
boto3==1.34.34
botocore==1.34.34

# Utils
python-dotenv==1.0.1
requests==2.31.0

# Testing
pytest==7.4.3

# Type Hints
typing-extensions==4.9.0

# HTTP
starlette==0.36.3
